Sarcasm detection
A very basic neural network to analyse 
The python program sarcasm.py is able to generate new models, which will replace the given one.
The simpleSarcasm.keras file is an example model. the playground allows you to test out this model on any sentence.
The dataset is provided for transparency purposes, and is not required to run the model. TensorFlow is required to run, along with its dependencies.