List to morphemes
Created a module that turns a list of words into a list of the individual morphemes that make up the word, given a list of morphemes to work from. Morpheme list should have elements in the form of {token : root}
Provided a list of morphemes, morphemes.json, and a function to parse them. Morphemes.json is Copyright (c) 2019 Colin Goldberg provided with all rights set out in the License also attached.